pyproject.toml
[project]
name = "voting"
version = "0.1.0"
description = "A secure voting system using Django"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url ~= 3.0",
    "dj-email-url ~= 1.0",
    "django ~= 5.2",
    "django-allauth[socialaccount] ~= 65.11",
    "django-countries ~= 7.6",
    "django-debug-toolbar ~= 6.0",
    "django-polymorphic ~= 4.1",
    "pgpy13 ~= 0.6.1rc1",
]

[tool.setuptools]
py-modules = []
